requirements.txt
pyqtgraph
numpy
pyqt5
i2cylib
ch347api
matplotlib
pyserial
